move UIConf
